pyproject.toml
[project]
name = "secret-santa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "phonenumbers>=9.0.21",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
]
